[e621] move e621 to a separate file (again)
