+++ gallery_dl/extractor/__init__.py
@@ -34,6 +34,7 @@ modules = [
     "desktopography",
     "deviantart",
     "dynastyscans",
+    "e621",
     "erome",
     "exhentai",
     "fallenangels",
